Fix compiler warning.

+++ src/iptc.cpp
@@ -183,7 +183,7 @@ namespace Exiv2 {
 
     long Iptcdatum::toLong(long n) const
     {
-        return value_.get() == nullptr ? -1 : value_->toInt64(n);
+        return value_.get() == nullptr ? -1 : static_cast<long>(value_->toInt64(n));
     }
 
     float Iptcdatum::toFloat(long n) const
